requirements.txt
Flask==2.0.1
gunicorn==20.1.0
pdf2image==1.16.0
pytesseract==0.3.8
spacy==3.1.3
python-docx==0.8.11
Pillow==8.3.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
python-dotenv==0.19.0
psycopg2-binary==2.9.1  # For PostgreSQL
bcrypt==3.2.0  # For password hashing
# ... any other dependencies you're using 